Use volume resource for ebs recipe
I had to move format_directory to environment libraries because
explicitly used by the ebs recipe.

###
* Moved tests for format_directory method.
