--- cookbooks/aws-parallelcluster-common/libraries/helpers.rb
@@ -1,5 +1,0 @@
-def format_directory(dir)
-  format_dir = dir.strip
-  format_dir = "/#{format_dir}" unless format_dir.start_with?('/')
-  format_dir
-end
