+++ cookbooks/aws-parallelcluster-environment/recipes/config/ebs.rb
@@ -27,24 +27,12 @@ when 'ComputeFleet'
 
   # Mount each volume with NFS
   shared_dir_array.each do |dir|
-    dirname = format_directory(dir)
-    exported_dirname = format_directory(dir)
-
-    # Created shared mount point
-    directory dirname do
-      mode '1777'
-      owner 'root'
-      group 'root'
-      recursive true
-      action :create
-    end
-
-    # Mount shared volume over NFS
-    mount dirname do
-      device(lazy { "#{node['cluster']['head_node_private_ip']}:#{exported_dirname}" })
+    volume "mount volume #{index}" do
+      action :mount
+      shared_dir dir
+      device(lazy { "#{node['cluster']['head_node_private_ip']}:#{format_directory(dir)}" })
       fstype 'nfs'
       options node['cluster']['nfs']['hard_mount_options']
-      action %i(mount enable)
       retries 10
       retry_delay 6
     end
