+++ cookbooks/aws-parallelcluster-environment/resources/volume.rb
@@ -130,9 +130,3 @@ action :unexport do
     command "exportfs -ra"
   end
 end
-
-def format_directory(dir)
-  format_dir = dir.strip
-  format_dir = "/#{format_dir}" unless format_dir.start_with?('/')
-  format_dir
-end
