+++ cookbooks/aws-parallelcluster-environment/spec/unit/resources/volume_spec.rb
@@ -1,33 +1,5 @@
 require 'spec_helper'
 
-describe 'volume:format_directory' do
-  cached(:chef_run) do
-    ChefSpec::SoloRunner.new.converge_dsl do
-      volume 'nothing' do
-        action :nothing
-      end
-    end
-  end
-
-  cached(:resource) { chef_run.find_resource('volume', 'nothing') }
-
-  it 'calls volume:nothing' do
-    is_expected.to nothing_volume('nothing')
-  end
-
-  it 'adds / at the beginning if not starting with /' do
-    expect(resource.format_directory('any')).to eq('/any')
-  end
-
-  it 'does not add / at the beginning if already starting with /' do
-    expect(resource.format_directory('/any')).to eq('/any')
-  end
-
-  it 'strips whitespaces from directory' do
-    expect(resource.format_directory('  any ')).to eq('/any')
-  end
-end
-
 describe 'volume:mount' do
   for_all_oses do |platform, version|
     context "on #{platform}#{version}" do
